Separate markdown parse out into own module for reuse.
